Move remaining tests into __tests__
